requirements.txt
coverage==4.4.1
dateutils==0.6.6
pyflakes==1.6.0
pep8==1.7.0
factory-boy==2.9.2
pytest==3.7.1
pytest-cov==2.5.1
